+++ db4e/Modules/DeplClient.py
@@ -84,6 +84,13 @@ class DeplClient:
         # Check for duplicate instance names and missing fields
         self.check_instance_and_fields(elem)
 
+        # Check for errors
+        if elem.status() != DStatus.GOOD and elem.status() != DStatus.UNKNOWN:
+            print(f"DeplClient:add_deployment(): Add failed: {elem.status()}")
+            for msg in elem.get_msgs():
+                print(f"DeplClient:add_deployment(): {msg}")
+            return elem
+
         class_map = {
             Db4E: DElem.DB4E,
             MoneroD: DElem.MONEROD,
@@ -95,6 +102,8 @@ class DeplClient:
 
         # Create an add job
         elem_type = class_map[type(elem)]
+
+        print(f"DeplClient:add_deployment(): Posting {DJob.NEW} job for {elem}")
         
         job = Job(op=DJob.NEW, instance=elem.instance(), elem_type=elem_type, elem=elem)
         self.job_queue.post_job(job)