+++ db4e/Panes/XMRigHashratesPane.py
@@ -13,7 +13,7 @@ from textual.widgets import Label, Select
 
 from db4e.Modules.XMRig import XMRig
 
-from db4e.Widgets.HashratePlot import HashratePlot
+from db4e.Widgets.Db4EPlot import Db4EPlot
 
 from db4e.Constants.DForm import DForm
 from db4e.Constants.DField import DField
@@ -28,7 +28,7 @@ class XMRigHashratesPane(Container):
     instance_label = Label("", id=DForm.INSTANCE_LABEL, classes=DForm.STATIC)
     hashrate_label = Label("", id=DForm.HASHRATE_LABEL, classes=DForm.STATIC)
     uptime_label = Label("", id=DForm.UPTIME_LABEL, classes=DForm.STATIC)
-    hashrate_plot = HashratePlot(DLabel.HASHRATE, id=DField.HASHRATE_PLOT)
+    hashrate_plot = Db4EPlot(DLabel.HASHRATE, id=DField.HASHRATE_PLOT)
     select_widget = Select(compact=True, id=DForm.TIMES, options=DSelect.HOURS_SELECT_LIST)
 
 
@@ -79,6 +79,6 @@ class XMRigHashratesPane(Container):
             hashrates = data[DField.VALUES]
             units = data[DField.UNITS]
             
-            plot = self.query_one("#" + DField.HASHRATE_PLOT, HashratePlot)
-            plot.load_data(days=days, hashrates=hashrates, units=units)
-            plot.hashrate_plot()  
+            plot = self.query_one("#" + DField.HASHRATE_PLOT, Db4EPlot)
+            plot.load_data(days=days, values=hashrates, units=units)
+            plot.db4e_plot()