+++ db4e/server.py
@@ -263,7 +263,6 @@ class Db4eServer:
                 thread.join()
         elif type(elem) == P2PoolRemote or type(elem) == MoneroDRemote:
             self.disable_downstream(elem)
-            elem.enabled(False)
             self.depl_mgr.delete_deployment(elem)
         elif type(elem) == MoneroD:
             self.disable_downstream(elem)
@@ -395,6 +394,7 @@ class Db4eServer:
         rc = sd.start()
         if rc == 0:
             self.log.info(f'Started: {elem}')
+            time.sleep(30)
         else:
             self.log.critical(f'ERROR: Failed to start {elem}, return code was {rc}')
             self.stopping.discard(instance)
@@ -432,6 +432,7 @@ class Db4eServer:
         rc = sd.stop()
         if rc == 0:
             self.log.info(f'Stopped: {elem}')
+            time.sleep(30)
             if isinstance(elem, P2Pool):
                 control = self.log_watchers.pop(instance, None)
                 if control:
@@ -607,9 +608,10 @@ class Db4eServer:
 
                 for elem_type, instance in depls:
                     if depls[(elem_type, instance)]:
-                        sd = self.systemd
-                        sd.service_name(elem_type + '@' + instance)
-                        sd.restart()
+                        # Create a restart job
+                        job = Job(op=DJob.RESTART, elem_type=elem_type, instance=instance)
+                        self.job_queue.post_job(job)
+
 
             except Exception as e:
                 self.log.error(f"rotate_logs(): {e} {stderr}")
@@ -739,6 +741,12 @@ class Db4eServer:
                 msgs += val[DField.MESSAGE] + "\n"
         job.msg(msgs[:-1])
         self.job_queue.complete_job(job)
+
+        # Restart Monerod and P2Pool deployments if their config has been updated
+        if type(elem) == MoneroD or type(elem) == P2Pool:
+            # Create a restart job 
+            job = Job(op=DJob.RESTART, elem_type=elem.elem_type(), instance=elem.instance())
+            self.job_queue.post_job(job)
 
 
     def unset_int_p2pool_primary(self):